Avoided redistributions at the end of the simulation by adding a check to ensure GVT was not TIME_INFINITY in MultiThreadedSimulation.cpp. If this check is not in place then there could be unprocessed events in shared multi-threading queues at the end of simulation.
A copule of minor compiler warning fixes in assertions in GVTMessage
and NumaMemoryManager.

+++ muse/kernel/src/NumaMemoryManager.cpp
@@ -278,7 +278,7 @@ NumaMemoryManager::redistribute(RedistributionMessage* rdm) {
 
 int
 NumaMemoryManager::getAllocatedMemory(const int numaID) const {
-    ASSERT((numaID >= 0) && (numaID < blockList.size()));
+    ASSERT((numaID >= 0) && (numaID < (int) blockList.size()));
     // Add the memory allocated -- each block is of fixed blockSize
     const int memAllocd = blockList[numaID].size() * blockSize -
         blockList[numaID].top().avail;
